feat: split out arena & vm state ops
